django-rbac/pyproject.toml
[tool.poetry]
name = "user-management-django"
version = "0.1.0"
description = ""
authors = ["KJD <125159447+Thrylos77@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
python-dotenv = "1.0.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.7.1"
django-simple-history = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
